+++ web/src/components/vis/LoadingsPlot.vue
