+++ web/src/components/vis/ScorePlot.vue
